test_setup: move some content to sub-modules
1. move `Setuper` and `SetupManager` to `test_setup.core`.

2. move `UlimitConfig` to `test_setup.os_posix`, which is introduced
   to place posix os related setup procedures.
